Add files via upload
This project explains about customer behavior patterns for an online eCommerce muti-category store. We can interpret data by seeing the customer buying and viewing behavior for a particular brand and category. The paper explains the method and process used for data manipulation and further analysis. The major goal of the project is to give a clear flow of handling big data files and data cleaning processes using Hadoop and Beeline. The data interpretation and analysis of this data was done using Excel, Tableau and Power BI. Charts and visuals like as bar diagrams, timeline and charts on customer behavior patterns for an online eCommerce muti-category store.